+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**?.?.? Changes:**
+
+* Added Heretic skins
+
 **1.0.0 Changes:**
 
 * First release